Code representing work done for 2nd year Data Structures and Algorithms module.
As stated in the readme this is simply meant as a demonstration of work I undertook in my second year of computer science at university in a data structures and algorithms module. This code implements some basic data structures/algorithms as required by the coursework on the module. Some code is unfinished, it is almost entirely uncommented and all of the code is bad! Implementations done to standards and requirements of coursework, not actual implementations!